+++ src/main/java/pl/engine/shapes/flat/Circle.java
@@ -1,6 +1,6 @@
 package pl.engine.shapes.flat;
 
-import pl.engine.Vector3;
+import pl.engine.math.Vector3;
 import pl.engine.shapes.Drawable;
 
 import java.awt.*;
@@ -8,7 +8,7 @@ import java.awt.*;
 public class Circle extends Drawable {
 
     private Vector3 mid;
-    private int r;
+    private double r;
 
     public Circle(Vector3 mid, int r, Color color){
 
@@ -20,15 +20,15 @@ public class Circle extends Drawable {
     @Override
     public void draw() {
 
-        int x;
-        int y;
+        double x;
+        double y;
 
         for(double deg = 0; deg <= 360; deg++){
 
             double degInRadians = Math.toRadians(deg);
 
-            x = (int) ((double) r * Math.cos(degInRadians)) + mid.x;
-            y = (int) ((double) r *  Math.sin(degInRadians)) + mid.y;
+            x = r * Math.cos(degInRadians) + mid.x;
+            y = r * Math.sin(degInRadians) + mid.y;
 
             drawPixel(x, y, color);
         }
