+++ src/main/java/pl/engine/shapes/flat/Disk.java
@@ -1,6 +1,6 @@
 package pl.engine.shapes.flat;
 
-import pl.engine.Vector3;
+import pl.engine.math.Vector3;
 import pl.engine.texture.Texturable;
 import pl.engine.texture.Texture;
 
@@ -10,7 +10,7 @@ import java.util.function.BiConsumer;
 public class Disk extends Texturable {
 
     private Vector3 mid;
-    private int r;
+    private double r;
 
     public Disk(Vector3 mid, int r, Color color){
         this(mid, r);
@@ -33,27 +33,27 @@ public class Disk extends Texturable {
     @Override
     public void draw() {
 
-        int y;
-        int prevY = Integer.MIN_VALUE;
+        double y;
+        double prevY = Integer.MIN_VALUE;
 
-        BiConsumer<Integer, Integer> draw = getDrawTextureOrColorPixelFunction();
+        BiConsumer<Double, Double> draw = getDrawTextureOrColorPixelFunction();
 
         for(double deg = 90; deg >= -90; deg--){
 
             double degInRadians = Math.toRadians(deg);
 
-            y = (int) Math.round((double) r * Math.sin(degInRadians)) + mid.y;
+            y = Math.round(r * Math.sin(degInRadians)) + mid.y;
 
-            int rightX = (int) ((double) r * Math.cos(degInRadians)) + mid.x;
-            int leftX = 2 * mid.x - rightX; // got from mid.x - (rightX - mid.x)
+            double rightX = r * Math.cos(degInRadians) + mid.x;
+            double leftX = 2 * mid.x - rightX; // got from mid.x - (rightX - mid.x)
 
-            int yDiff = prevY - y;
+            double yDiff = prevY - y;
 
             if(prevY != Integer.MIN_VALUE && yDiff > 1){
 
-                for(int missingY = prevY - 1; missingY >= y; missingY--){
+                for(double missingY = prevY - 1; missingY >= y; missingY--){
 
-                    for(int x = leftX; x <= rightX; x++){
+                    for(double x = leftX; x <= rightX; x++){
 
                         draw.accept(x, missingY);
                     }
@@ -61,7 +61,7 @@ public class Disk extends Texturable {
             }
             else {
 
-                for(int x = leftX; x <= rightX; x++){
+                for(double x = leftX; x <= rightX; x++){
 
                     draw.accept(x, y);
                 }
